Use dependency attributes when selecting variants of a single component
This commit makes sure that attributes defined on a dependency are used whenever
a component has multiple variants, and we need to select one of the variants.
However, this doesn't work yet when an attribute on a dependency overrides the
consumer attributes (the attributes defined on a configuration). In this case,
we will rightfully select the variant, but as soon as we try to get the artifacts
of this variant, we will fail, because we're selecting _again_ the variant, but
the only thing we have at this point is the selected component and the consumer
attributes (we have lost the dependency attributes).